pyproject.toml
[project]
name = "vec_search_util"
version = "0.1"
requires-python = ">=3.10"
license = {file = "LICENSE"}

dependencies = [
    "transformers==4.36.2",
    "sqlite-vss==0.1.2",
    "open_clip_torch==2.23.0",
    "pandas",
    "scikit-learn"
]

#[project.optional-dependencies]
#dev = []


[tool.setuptools]
# ...
# By default, include-package-data is true in pyproject.toml, so you do
# NOT have to specify this line.
include-package-data = true

#[tool.setuptools.packages.find]
#where = ["vsu"]

#[tool.setuptools.package-data]
#sample = ["*.csv", "img/*.png"]
